perf: optimize api requests and add redis for storing session data
Make changes to getOnePokemon to perform MongoDB search instead of fetching all data and filtering it on the server.

Make changes to pokédex and capture pages to utilize MongoDB sort instead of sorting on the server.

Switch to storing session data in a Redis database instead of in MongoDB.